refactor(ai-plan-import): isole le pipeline d'extraction dans pipeline/
Deplace run-import-pipeline, extract-text et les etapes (extract-actions,
enrich-sous-actions, consolidate-actions, qualitative-review, score-actions)
hors de generate-import-draft, qui ne conserve que le worker, le service et
le mapper draft vers plan. Aucun changement de comportement.
