+++ apps/backend/src/plans/plans/ai-plan-import/pipeline/__fixtures__/sample.pdf
