+++ apps/backend/src/plans/plans/ai-plan-import/pipeline/score-actions/score-actions.schema.ts
